# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduBridge Egypt — Education platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      parameters:
        - name: teacherId
          in: query
          schema:
            type: integer
        - name: subject
          in: query
          schema:
            type: string
        - name: gradeLevel
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Courses list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /courses/{courseId}/enroll:
    post:
      operationId: enrollInCourse
      tags: [courses]
      summary: Enroll student in course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollBody"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /courses/{courseId}/students:
    get:
      operationId: getCourseStudents
      tags: [courses]
      summary: List students enrolled in a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentEnrollment"

  # ── LESSONS ───────────────────────────────────────────────────────────────
...
  /courses/{courseId}/announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements in a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  # ── MESSAGES ──────────────────────────────────────────────────────────────
...
  /messages/conversations/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/start:
    post:
      operationId: startConversation
      tags: [messages]
      summary: Start a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationBody"
      responses:
        "201":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  # ── PAYMENTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: string
        teacherAvatarUrl:
          type: string
        subject:
          type: string
        gradeLevel:
          type: string
        price:
          type: number
        currency:
          type: string
          default: EGP
        coverImageUrl:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        enrollmentCount:
          type: integer
        rating:
          type: number
        lessonCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, teacherId, subject, status, createdAt]
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [courses, total, page, limit]
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        subject:
          type: string
        gradeLevel:
          type: string
        price:
          type: number
        coverImageUrl:
          type: string
      required: [title, subject]
    UpdateCourseBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        subject:
          type: string
        gradeLevel:
          type: string
        price:
          type: number
        coverImageUrl:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        course:
          $ref: "#/components/schemas/Course"
        enrolledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, completed, dropped]
        progress:
          type: number
      required: [id, studentId, courseId, enrolledAt, status]
    EnrollBody:
      type: object
      properties:
        studentId:
          type: integer
        paymentId:
          type: integer
      required: [studentId]
    StudentEnrollment:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        studentAvatarUrl:
          type: string
        enrolledAt:
          type: string
          format: date-time
        progress:
          type: number
        lastActive:
          type: string
          format: date-time
      required: [studentId, studentName, enrolledAt]

    # ── LESSON ──
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: string
        title:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, courseId, authorId, title, body, createdAt]
    CreateAnnouncementBody:
      type: object
      properties:
        authorId:
          type: integer
        title:
          type: string
        body:
          type: string
      required: [authorId, title, body]

    # ── MESSAGES ──
    Conversation:
      type: object
      properties:
        id:
          type: integer
        participantIds:
          type: array
          items:
            type: integer
        participantNames:
          type: array
          items:
            type: string
        participantAvatars:
          type: array
          items:
            type: string
        lastMessage:
          type: string
        lastMessageAt:
          type: string
          format: date-time
        unreadCount:
          type: integer
      required: [id, participantIds, participantNames, lastMessageAt, unreadCount]
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderAvatarUrl:
          type: string
        body:
          type: string
        sentAt:
          type: string
          format: date-time
        read:
          type: boolean
      required: [id, conversationId, senderId, body, sentAt, read]
    SendMessageBody:
      type: object
      properties:
        senderId:
          type: integer
        body:
          type: string
      required: [senderId, body]
    StartConversationBody:
      type: object
      properties:
        initiatorId:
          type: integer
        recipientId:
          type: integer
        firstMessage:
          type: string
      required: [initiatorId, recipientId, firstMessage]

    # ── PAYMENT ──
...